requirements.txt
opencv-python>=4.8.0
tensorflow>=2.16.0
torch>=2.2.0
torchvision>=0.17.0
mediapipe>=0.10.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pillow>=10.0.0
tqdm>=4.65.0
albumentations>=1.3.0
imgaug>=0.4.0
tensorboard>=2.13.0
seaborn>=0.12.0
pandas>=2.0.0
pycocotools>=2.0.0
tensorflow-model-optimization>=0.7.0
onnx>=1.14.0
onnxruntime>=1.15.0
